+++ src/Bernard/Producer.php
@@ -2,24 +2,25 @@
 
 namespace Bernard;
 
-use Bernard\Middleware\MiddlewareBuilder;
+use Bernard\Event\EnvelopeEvent;
+use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
 /**
  * @package Bernard
  */
-class Producer implements Middleware
+class Producer
 {
     protected $queues;
-    protected $middleware;
+    protected $dispatcher;
 
     /**
-     * @param QueueFactory      $queues
-     * @param MiddlewareBuilder $middleware
+     * @param QueueFactory    $queues
+     * @param EventDispatcherInterface $dispatcher
      */
-    public function __construct(QueueFactory $queues, MiddlewareBuilder $middleware)
+    public function __construct(QueueFactory $queues, EventDispatcherInterface $dispatcher)
     {
         $this->queues = $queues;
-        $this->middleware = $middleware;
+        $this->dispatcher = $dispatcher;
     }
 
     /**
@@ -28,17 +29,11 @@ class Producer implements Middleware
      */
     public function produce(Message $message, $queueName = null)
     {
-        $queue = $this->queues->create($queueName ?: bernard_guess_queue($message));
+        $queueName = $queueName ?: bernard_guess_queue($message);
 
-        $middleware = $this->middleware->build($this);
-        $middleware->call(new Envelope($message), $queue);
-    }
+        $queue = $this->queues->create($queueName);
+        $queue->enqueue($envelope = new Envelope($message));
 
-    /**
-     * {@inheritDoc}
-     */
-    public function call(Envelope $envelope, Queue $queue)
-    {
-        $queue->enqueue($envelope);
+        $this->dispatcher->dispatch('bernard.produce', new EnvelopeEvent($envelope, $queue));
     }
 }