+++ tests/Bernard/Tests/ConsumerTest.php
@@ -7,15 +7,49 @@ use Bernard\Queue\InMemoryQueue;
 use Bernard\Envelope;
 use Bernard\Message\DefaultMessage;
 use Bernard\Router\SimpleRouter;
-use Bernard\Middleware\MiddlewareBuilder;
+use Bernard\Event\RejectEnvelopeEvent;
+use Bernard\Event\EnvelopeEvent;
 
 class ConsumerTest extends \PHPUnit_Framework_TestCase
 {
     public function setUp()
     {
         $this->router = new SimpleRouter;
-        $this->middleware = new MiddlewareBuilder;
-        $this->consumer = new Consumer($this->router, $this->middleware);
+        $this->router->add('ImportUsers', new Fixtures\Service);
+
+        $this->dispatcher = $this->getMock('Symfony\Component\EventDispatcher\EventDispatcherInterface');
+        $this->consumer = new Consumer($this->router, $this->dispatcher);
+    }
+
+    public function testEmitsConsumeEvent()
+    {
+        $envelope = new Envelope(new DefaultMessage('ImportUsers'));
+        $queue = new InMemoryQueue('queue');
+
+        $this->dispatcher->expects($this->at(0))->method('dispatch')
+            ->with('bernard.invoke', new EnvelopeEvent($envelope, $queue));
+
+        $this->dispatcher->expects($this->at(1))->method('dispatch')
+            ->with('bernard.acknowledge', new EnvelopeEvent($envelope, $queue));
+
+        $this->consumer->invoke($envelope, $queue);
+    }
+
+    public function testEmitsExceptionEvent()
+    {
+        $exception = new \InvalidArgumentException();
+
+        $this->router->add('ImportUsers', function () use ($exception) {
+            throw $exception;
+        });
+
+        $envelope = new Envelope(new DefaultMessage('ImportUsers'));
+        $queue = new InMemoryQueue('queue');
+
+        $this->dispatcher->expects($this->at(1))->method('dispatch')
+            ->with('bernard.reject', new RejectEnvelopeEvent($envelope, $queue, $exception));
+
+        $this->consumer->invoke($envelope, $queue);
     }
 
     public function testShutdown()